New Stuff/Controls
!+ includes/cub3.h
   ^^ changed from int to take a double.

!+ includes/struct.h
   ^^ Changed so that every texture has it's own image structure.
   ^^ Changed the player's structure. From int, to double.

/| src/init.c
   ^^ Created some basic controls so that the player can rotate on it's axis. STILL IN DEVELOPMENT.
   ^^ In Load Images, I assign every image structure.
   ^^ In Init_data(), I made some small readjustments.

!+ src/init_utils2.c
   ^^ changed from int to take a double.

+++ includes/cub3d.h
@@ -60,5 +60,5 @@ int open_fd(char *argv);
 
 void	get_colors(char *argv, char ***floor_colors, char ***ceiling_colors);
 void	get_window_size(char *argv, size_t	*height, size_t	*width);
-void	get_player_position(char	***map, int *x, int *y);
+void	get_player_position(char	***map, double *x, double *y);
 #endif
